This is a getting started doc
This doc takes the current tutorial doc and updates it. The current git-repo google skaffold no longer holds the proper resources in the repository for the tutorial. Instead of editing the tutorial doc. I made a copy and called it getting-started.md. I edited this copy to hold relevant resources. I created a copy instead of editing because tutorial.md is referenced in many places and was not sure about the effects of editing/renaming. This getting-started shows the user how to make a simple hello world task. The tutorial goes further with the hello world. Showing the user how to create an entire pipeline that will build an image from source code, push it to a container registry, and deploy in onto the cluster. There were changes made for the user to use dockerhub instead of a local docker registry. This tutorial has also been manually tested on a minikube cluster.